added loading data from API server into listView

+++ app/src/main/java/com/maskjs/korona_zakupy/data/orders/UsersInfoModel.kt
@@ -6,7 +6,7 @@ data class UsersInfoModel(
     val name:String,
     val firstName:String,
     val lastName:String,
-    val rating: Long,
+    val rating: Double,
     val photoDirectory:String?
 ) {
 }
